+++ Day 7/part2.c
@@ -13,8 +13,8 @@ int convert_str_to_arr(char *str, long *nums);
 int main() {  
   char buffer[BUFFER_SIZE];
   FILE *f = fopen("./input.txt", "r");
-
-  unsigned long long totalCalibrationResult = 0;
+  
+  long totalCalibrationResult = 0;
   while(fgets(buffer, BUFFER_SIZE, f)) {
     char bufferCopy[BUFFER_SIZE];
     strcpy(bufferCopy, buffer);
@@ -35,7 +35,7 @@ int main() {
   }  
   fclose(f);
 
-  printf("Total Calibration Result: %lld\n", totalCalibrationResult);
+  printf("Total Calibration Result: %ld\n", totalCalibrationResult);
   return 0;
 }
